+++ lib/legion/extensions/dynatrace/ownership.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative 'ownership/runners/ownership'
+require_relative 'helpers/client'
+require_relative 'base_client'
+
+module Legion
+  module Extensions
+    module Dynatrace
+      module Ownership
+        class Client < BaseClient
+          include Runners::Ownership
+        end
+      end
+    end
+  end
+end